Do not append to shared slice
If two calls to this function happen at ~the same time, one of them will
append to the shared c.Tags slice, and then the other one will. If
c.Tags has enough room to avoid being reallocated, then it will be
mutated directly instead, and the second append will overwrite the first
one. We should not be appending to a slice that is scoped to the entire
statsd client.