move statsd_test.go to statsd_test to force all the existing tests to ...
rely on the interface instead of the implementation.
I removed all the tests that were testing the implementation.
The main goal here is to detach the existing unit tests from the implementation
to be able to change the implementation and still test it with the old unit tests.

A few benchmarks that were not useful were also removed at the same time.