statsd: close the stop channel when closing a statsd client to avoid leaking. (#107)
This patch fixes a resource leak that occurs when closing a statsd client.
The Close function sent an empty struct to the stop channel, but there were two
goroutines waiting on the stop channel. Since the Close function did not close
the stop channel, one of the two goroutines would not exit, and would instead be
leaked along with any memory resources the statsd client held.

This patch causes the Close function to close the stop channel, ensuring any
goroutines listening on that channel will know the client has been stopped.