[Beta] Adding client side aggregation to the client
This should reduce by a lot the number of packets (and therefore the
number of packets drop) between the client and the Agent. This should
also improve performances in hot path when sampling the same metrics.

Using fnv1a hash for the aggregator gives us a average ~12% improvement
on the aggregation benchmarks.