Implement ChannelMode and sampling rate for extended aggregation
When using extended aggregation we still want to respect sampling rate
for histograms, distribution and timing as it will have direct
consequences on the Agent (number of point to aggregate).

For apps sending a high number of metrics the aggregator implements
ChannelMode to avoid lock contention when generating random numbers.