Fix telemetry with extended aggregation and channelMode (#194)
When extended aggregation is enabled with channelMode we don't need to
run the workers in channelMode since the user never use them directly
(the aggregator sit in between). The telemetry on the other side was
using the workers in channelMode despite them not being reading on the
end of the channel. This caused the telemetry to never been sent.